# series solution