--- a/OpenMP2/src/main.cpp
+++ b/OpenMP2/src/main.cpp
@@ -2,10 +2,30 @@
 #include <iomanip>
 #include <fstream>
 #include <omp.h>
+#include <string>
 
 double calc(uint32_t x_last, uint32_t num_threads)
 {
-  return 0;
+  double res(0.0);
+  double* token = new double[x_last];
+
+  //  [0]  [1]  [2]
+  // sum0 sum1 sum2
+  
+  #pragma omp parallel num_threads(num_threads)
+  { 
+    #pragma omp for
+    for (int i = x_last; i > 0; i--)
+    {
+      token[i - 1] += (1.0) / i;
+    }
+  }
+
+  for(int i(x_last - 1); i >= 0; i--)
+    res += token[i];
+  
+  delete[] token;
+  return res;
 }
 
 int main(int argc, char** argv)