--- a/OpenMP3/src/main.cpp
+++ b/OpenMP3/src/main.cpp
@@ -16,10 +16,11 @@ double calc(double x0, double x1, double dx, uint32_t num_threads)
   double sum(0.0);
 
   // allocate store buffer
+  // buffer size is significantly less than x1-x0 to avoid frequents RAM access
 
   double* token = new double[TOKEN_NUM];
 
-  int i(0);
+  // int i(0);
   int num_steps = (x1 - x0) / dx; // 11
   // std::cout << "num steps " << num_steps << std::endl;
   int token_size = num_steps / TOKEN_NUM; // 5
@@ -28,24 +29,22 @@ double calc(double x0, double x1, double dx, uint32_t num_threads)
   #pragma omp parallel num_threads(num_threads)
   {
     #pragma omp for
-    for(int i = (0); i < token_size * TOKEN_NUM; ++i)
+    for(int i = (1); i < token_size * TOKEN_NUM; ++i)
     {
       int token_idx = i / token_size; // 0 0 0 0 0 1 1 1 1 1 1 2
-      token[token_idx] += (func(x0 + i * dx) + func(x0 + (i + 1) * dx));
-      // sum += (func(x0 + i * dx) + func(x0 + (i + 1) * dx));
+      token[token_idx] += (func(x0 + i * dx));
     }
   }
   for(int i(token_size * TOKEN_NUM); i < num_steps; ++i)
-    sum += (func(x0 + i * dx) + func(x0 + (i + 1) * dx));
+    sum += (func(x0 + i * dx));// + func(x0 + (i + 1) * dx));
 
   for(int i(0); i < TOKEN_NUM; ++i)
   {
     sum += token[i];
     // std::cout << token[i] << std::endl;
   }
+  sum += (func(x0) + func(x1)) / 2.0;
   sum *= dx;
-  sum /= 2.0;
-
 
   delete token;
   return sum;