--- a/OpenMP4/src/main.cpp
+++ b/OpenMP4/src/main.cpp
@@ -6,7 +6,7 @@
 
 enum BORDERS
 {
- BORDER_DEN = 100,
+ BORDER_DEN = 10,
  BORDER_NUM = 2
 };
 
@@ -18,6 +18,7 @@ double inverse_fact(int n)
   {
     fact /= double(i);
   }
+  // std::cout << "inverse fact[" << n << "] " << fact << std::endl;
   return fact;
 }
 
@@ -27,65 +28,69 @@ double inverse_fact_from_border(int n, double border_value, int BORDER)
   for(int i(BORDER + 1); i <= n; ++i)
   {
     fact /= double(i);
+    // std::cout << "inverse fact from border: " << fact << std::endl;
   }
+  // std::cout << "border value: " << border_value << std::endl;
+  // std::cout << "inverse fact from border: " << fact * border_value << std::endl;
   return fact * border_value;
 }
 
-double calc(uint32_t x_last, uint32_t num_threads)
+double calc(int x_last, int num_threads)
 {
+  std::cout << "STEPS: " << x_last << std::endl;
+	num_threads++;
+	num_threads--;
  /* int BORDER = x_last / BORDER_DEN;
   int BORDER1 = x_last / 10;*/
   double sum(0.0);
   // double border_value(0.0);
 
   int* borders = new int[BORDER_NUM];
-  int* border_value = new int[BORDER_NUM];
+  double* border_value = new double[BORDER_NUM];
   int border_size = x_last / BORDER_NUM;
   for(int i(0); i < BORDER_NUM; ++i)
   {
     borders[i] = i * border_size;
-    std::cout << "borders[" << i << "] = " << borders[i] << std::endl;
+    // std::cout << "borders[" << i << "] = " << borders[i] << std::endl;
   }
- 
-  for(int i(0); i < x_last; ++i)
+
+  double* res = new double[num_threads];
+  #pragma omp parallel num_threads(num_threads)
   {
-    double value(0.0);
-
-    for(int j(1); j < BORDER_NUM - 1; ++j)
+    #pragma omp for
+    for(int i = 0; i < x_last; ++i)
     {
-      if(i > borders[j + 1])
-        value = inverse_fact_from_border(i, border_value[j + 1], borders[j + 1]);
-      if((i <= borders[j + 1]) && (i > borders[j]))
-        value = inverse_fact_from_border(i, border_value[j], borders[j]);
-      if(i == borders[j])
-        border_value[j] = value;
-    }
-    if(i <= borders[1])
+      double value(0.0);
+      if(i <= borders[1])
         value = inverse_fact(i);
-    sum += value;
 
-   /* if(i > BORDER1)
-      value = inverse_fact_from_border(i, border_value, BORDER1);
-    if((i > BORDER) && (i <= BORDER1))
-      value = inverse_fact_from_border(i, border_value, BORDER);
-    if(i <= BORDER)
-      value = inverse_fact(i);
-    sum += value;
-    if(i == BORDER)
-    {
-      border_value = value;
-      std::cout << "border value: " << 1.0 / border_value << std::endl;
+      for(int j = 1; j < BORDER_NUM; ++j)
+      {
+        /*if(i > borders[j + 1])
+          value = inverse_fact_from_border(i, border_value[j + 1], borders[j + 1]);
+        if((i <= borders[j + 1]) && (i > borders[j]))
+          value = inverse_fact_from_border(i, border_value[j], borders[j]);*/
+        if(i == borders[j])
+        {
+          border_value[j] = value;
+          // std::cout << " i = " << i << " border value[" << j << "] = " << border_value[j] << std::endl;
+        } else if(i > borders[j])
+          value = inverse_fact_from_border(i, border_value[j], borders[j]);
+      }
+      int tid = omp_get_thread_num();
+      res[tid] += value;
+      // sum += value;
     }
-    if(i == BORDER1)
-    {
-      border_value = value;
-      std::cout << "border value1: " << 1.0 / border_value << std::endl;
-    }*/
   }
+  sum = 0.0;
+  for(int i(0); i < num_threads; ++i)
+    sum += res[i];
 
   delete[] borders;
   delete[] border_value;
+  delete[] res;
   
+  // std::cout << "SUM = " << sum << std::endl;
   return sum;
 }
 
@@ -116,7 +121,7 @@ int main(int argc, char** argv)
   }
 
 // Read arguments from input
-  uint32_t x_last = 0, num_threads = 0;
+  int x_last = 0, num_threads = 0;
   input >> x_last >> num_threads;
 
   // Calculation