[TEST 00] OK (2 ms)

[TEST 01]
  ./build/task ./tests/01/input.txt ./build/01.txt
[TEST 01] OK (2 ms)

[TEST 02]
  ./build/task ./tests/02/input.txt ./build/02.txt
[TEST 02] OK (5 ms)

[TEST 03]
  ./build/task ./tests/03/input.txt ./build/03.txt
[TEST 03] OK (1976 ms)

[TEST 04]
  ./build/task ./tests/04/input.txt ./build/04.txt
[TEST 04] OK (1088 ms)
===========
SUCCESSFUL: 00 01 02 03 04
FAIL: