@SokolovVadim 2020-10-08
Speed up calculations by caching values of factorial
[TEST 00]
  ./build/task ./tests/00/input.txt ./build/00.txt
borders[0] = 0
borders[1] = 0
[TEST 00] OK (2 ms)

[TEST 01]
  ./build/task ./tests/01/input.txt ./build/01.txt
borders[0] = 0
borders[1] = 0
[TEST 01] OK (2 ms)

[TEST 02]
  ./build/task ./tests/02/input.txt ./build/02.txt
borders[0] = 0
borders[1] = 1
[TEST 02] OK (3 ms)

[TEST 03]
  ./build/task ./tests/03/input.txt ./build/03.txt
borders[0] = 0
borders[1] = 1
[TEST 03] OK (2 ms)

[TEST 04]
  ./build/task ./tests/04/input.txt ./build/04.txt
borders[0] = 0
borders[1] = 5000
[TEST 04] OK (100 ms)

[TEST 05]
  ./build/task ./tests/05/input.txt ./build/05.txt
borders[0] = 0
borders[1] = 5000
[TEST 05] OK (101 ms)

[TEST 06]
  ./build/task ./tests/06/input.txt ./build/06.txt
borders[0] = 0
borders[1] = 150000
[TEST 06] OK (84952 ms)