@SokolovVadim 2020-10-14
# OpenMP3 performance increase:
## Previous results:

[TEST 00]
  ./build/task ./tests/00/input.txt ./build/00.txt
[TEST 00] OK (2 ms)

[TEST 01]
  ./build/task ./tests/01/input.txt ./build/01.txt
[TEST 01] OK (3 ms)

[TEST 02]
  ./build/task ./tests/02/input.txt ./build/02.txt
[TEST 02] OK (2488 ms)

[TEST 03]
  ./build/task ./tests/03/input.txt ./build/03.txt
[TEST 03] OK (719 ms)

[TEST 04]
  ./build/task ./tests/04/input.txt ./build/04.txt
[TEST 04] OK (3059 ms)

[TEST 05]
  ./build/task ./tests/05/input.txt ./build/05.txt
[TEST 05] OK (913 ms)

[TEST 06]
  ./build/task ./tests/06/input.txt ./build/06.txt
[TEST 06] OK (28274 ms)

[TEST 07]
  ./build/task ./tests/07/input.txt ./build/07.txt
[TEST 07] OK (7710 ms)

SUCCESSFUL: 00 01 02 03 04 05 06 07
FAIL: 

## New results
[TEST 00]
  ./build/task ./tests/00/input.txt ./build/00.txt
[TEST 00] OK (2 ms)

[TEST 01]
  ./build/task ./tests/01/input.txt ./build/01.txt
[TEST 01] OK (3 ms)

[TEST 02]
  ./build/task ./tests/02/input.txt ./build/02.txt
[TEST 02] OK (1229 ms)

[TEST 03]
  ./build/task ./tests/03/input.txt ./build/03.txt
[TEST 03] OK (387 ms)

[TEST 04]
  ./build/task ./tests/04/input.txt ./build/04.txt
[TEST 04] OK (1532 ms)

[TEST 05]
  ./build/task ./tests/05/input.txt ./build/05.txt
[TEST 05] OK (466 ms)

[TEST 06]
  ./build/task ./tests/06/input.txt ./build/06.txt
[TEST 06] OK (14206 ms)

[TEST 07]
  ./build/task ./tests/07/input.txt ./build/07.txt
[TEST 07] OK (4123 ms)

SUCCESSFUL: 00 01 02 03 04 05 06 07
FAIL: 

### So performance has significantly improved