@SokolovVadim 2020-10-14
# OpenMP4 Parallel processing

The performance has increased due to usage of `omp for` directive.
But the point is that `test 5` has failed.
Can't resolve this issue because output test value differs from exact value

[TEST 00]
  ./build/task ./tests/00/input.txt ./build/00.txt
STEPS: 1
[TEST 00] OK (2 ms)

[TEST 01]
  ./build/task ./tests/01/input.txt ./build/01.txt
STEPS: 1
[TEST 01] OK (3 ms)

[TEST 02]
  ./build/task ./tests/02/input.txt ./build/02.txt
STEPS: 2
[TEST 02] OK (3 ms)

[TEST 03]
  ./build/task ./tests/03/input.txt ./build/03.txt
STEPS: 2
[TEST 03] OK (3 ms)

[TEST 04]
  ./build/task ./tests/04/input.txt ./build/04.txt
STEPS: 24
[TEST 04] OK (4 ms)

[TEST 05]
  ./build/task ./tests/05/input.txt ./build/05.txt
STEPS: 24
[TEST 05] DIFF FAIL(3 ms): vimdiff ./build/05.txt ./tests/05/output.txt

[TEST 06]
  ./build/task ./tests/06/input.txt ./build/06.txt
STEPS: 50
[TEST 06] OK (2 ms)

[TEST 07]
  ./build/task ./tests/07/input.txt ./build/07.txt
STEPS: 50
[TEST 07] OK (3 ms)

[TEST 08]
  ./build/task ./tests/08/input.txt ./build/08.txt
STEPS: 10000000
[TEST 08] OK (1897123 ms)

[TEST 09]
  ./build/task ./tests/09/input.txt ./build/09.txt
STEPS: 10000000
[TEST 09] OK (495323 ms)

SUCCESSFUL: 00 01 02 03 04 06 07 08 09
FAIL: 05